% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsib.R
\name{catsib}
\alias{catsib}
\title{CATSIB DIF detection procedure}
\usage{
catsib(
  x = NULL,
  data,
  score = NULL,
  se = NULL,
  group,
  focal.name,
  D = 1,
  n.bin = c(80, 10),
  min.binsize = 3,
  max.del = 0.075,
  alpha = 0.05,
  missing = NA,
  purify = FALSE,
  max.iter = 10,
  min.resp = NULL,
  method = "MLE",
  range = c(-4, 4),
  norm.prior = c(0, 1),
  nquad = 41,
  weights = NULL,
  ncore = 1,
  verbose = TRUE
)
}
\arguments{
\item{x}{A data frame containing the item metadata (e.g., item parameters, number of categories, models ...). 
This metadata is required to estimate latent ability parameters, expecially when \code{purify = TRUE}. Default is NULL. 
See \code{\link{est_irt}}, \code{\link{irtfit}}, \code{\link{test.info}} or \code{\link{simdat}} for more details about the item metadata.}

\item{data}{A matrix containing examinees' response data for the items in the argument \code{x}. A row and column indicate
the examinees and items, respectively.}

\item{score}{A vector of examinees' ability estimates. If the abilities are not provided, \code{\link{catsib}} function estimates the abilities before 
computing RDIF statistics. See \code{\link{est_score}} for more details about scoring methods. Default is NULL.}

\item{se}{A vector of the standard errors of ability estimates. Default is NULL.}

\item{group}{A numeric or character vector indicating group membership of examinees. The length of vector should the same with the number of rows 
in the response data matrix.}

\item{focal.name}{A single numeric or character indicating the level of group which corresponds to the focal group. 
For example, if \code{group = c(0, 1, 0, 1, 1)} and '1' indicates the focal group, then \code{focal.name = 1}.}

\item{D}{A scaling factor in IRT models to make the logistic function as close as possible to the normal ogive function (if set to 1.7).
Default is 1.}

\item{n.bin}{}

\item{min.binsize}{}

\item{max.del}{}

\item{alpha}{A numeric value to specify significance \eqn{\alpha}-level of the hypothesis test using the RDIF fit statistics.
Default is .05.}

\item{missing}{A value indicating missing values in the response data set. Default is NA.}

\item{purify}{A logical value indicating whether a purification process will be implemented or not. Default is FALSE.}

\item{max.iter}{An integer value specifying a maximum number of iterations for the purification process. Default is 10.}

\item{min.resp}{An integer value?? Default is NULL.}

\item{method}{A character string indicating a scoring method. Available methods are "MLE" for the maximum likelihood estimation, 
"MAP" for the maximum a posteriori estimation, and "EAP" for the expected a posteriori estimation. Default method is "MLE".}

\item{range}{A numeric vector of two components to restrict the range of ability scale for the MLE. Default is c(-4, 4).}

\item{norm.prior}{A numeric vector of two components specifying a mean and standard deviation of the normal prior distribution.
These two parameters are used to obtain the gaussian quadrature points and the corresponding weights from the normal distribution. Default is
c(0,1). Ignored if \code{method} is "MLE".}

\item{nquad}{An integer value specifying the number of gaussian quadrature points from the normal prior distribution. Default is 41.
Ignored if \code{method} is "MLE" or "MAP".}

\item{weights}{A two-column matrix or data frame containing the quadrature points (in the first column) and the corresponding weights
(in the second column) of the latent variable prior distribution. The weights and quadrature points can be easily obtained
using the function \code{\link{gen.weight}}. If NULL and \code{method} is "EAP", default values are used (see the arguments
of \code{norm.prior} and \code{nquad}). Ignored if \code{method} is "MLE" or "MAP".}

\item{ncore}{The number of logical CPU cores to use. Default is 1. See \code{\link{est_score}} for details.}

\item{verbose}{A logical value. If TRUE, the progress messages of purification procedure are suppressed. Default is TRUE.#'}
}
\value{

}
\description{
This function analyze DIF on items using CATSIB method
}
\details{

}
\examples{


}
\references{
Lim, H., Edison, M. Choe., Han, K. T., Lee, S., & Hong, M. (2021, June). \emph{IRT residual approach 
to detecting DIF.} Paper presented at the Annual Meeting of the National Council on Measurement 
in Education. Online.
Nandakumar, R., & Roussos, L. (2004). Evaluation of the CATSIB DIF procedure in a pretest setting. 
\emph{Journal of Educational and Behavioral Statistics, 29}(2), 177-199.
}
\seealso{
\code{\link{rdif}}, \code{\link{est_item}}, \code{\link{test.info}}, \code{\link{simdat}}, 
\code{\link{shape_df}}, \code{\link{gen.weight}}, \code{\link{est_score}}
}
\author{
Hwanggyu Lim \email{hglim83@gmail.com}
}
