% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_score.R
\name{est_score}
\alias{est_score}
\title{Estimate examinees' ability (proficiency) parameters}
\usage{
est_score(x, data, D = 1, method = "MLE", range = c(-4, 4),
  norm.prior = c(0, 1), nquad = 41, weights = NULL, se = TRUE,
  missing = NA, ncore = 1, ...)
}
\arguments{
\item{x}{A data.frame containing the item meta data (e.g., item parameters, number of categories, models ...).
See \code{\link{irtfit}}, \code{\link{test.info}}, or \code{\link{simdat}} for more details about the item meta data.
This data.frame can be easily obtained using the function \code{\link{shape_df}}.}

\item{data}{A matrix containing examinees' response data for the items in the argument \code{x}. A row and column indicate
the examinees and items, respectively.}

\item{D}{A scaling factor in IRT models to make the logistic function as close as possible to the normal ogive function (if set to 1.7).
Default is 1.}

\item{method}{A character string indicating a scoring method. Available methods are "MLE" for
the maximum likelihood estimation, "MAP" for the maximum a posteriori estimation, "EAP" for the expected a posteriori estimation,
and "EAP.SUM" for the expected a posteriori summed scoring. Default method is "MLE".}

\item{range}{A numeric vector of two components to restrict the range of ability scale for the MLE. Default is c(-4, 4).}

\item{norm.prior}{A numeric vector of two components specifying a mean and standard deviation of the normal prior distribution.
These two parameters are used to obtain the gaussian quadrature points and the corresponding weights from the normal distribution. Default is
c(0,1). Ignored if \code{method} is "MLE".}

\item{nquad}{An integer value specifying the number of gaussian quadrature points from the normal prior distribution. Default is 41.
Ignored if \code{method} is "MLE" or "MAP".}

\item{weights}{A two-column matrix or data.frame containing the theta values (in the first column) and the weights (in the second column)
for the prior distribution. The weights and theta values can be easily obtained using the function \code{\link{gen.weight}}.
If NULL and \code{method} is "EAP" or "EAP.SUM", default values are used (see the arguments of \code{norm.prior} and \code{nquad}). Ignored
if \code{method} is "MLE" or "MAP".}

\item{se}{A logical value. If TRUE, the standard errors of ability estimates are computed. However, if \code{method} is "EAP.SUM", the standard
errors are always returned.}

\item{missing}{A value indicating missing values in the response data set. Default is NA. See below for details.}

\item{ncore}{The number of logical CPU cores to use. Default is 1. See below for details.}

\item{...}{additional arguments to pass to \code{parallel::makeCluster}.}
}
\value{
A list including a vector of the ability estimates and a vector of the standard errors of ability estimates. When "EAP.SUM" is used,
raw sum scores of examinees and a table with the possible raw sum scores and corresponding ability estimates are returned as well.
}
\description{
This function estimates examinees' latent ability parameters. Available scoring methods are maximum likelihood estimation (MLE),
maximum a posteriori estimation (MAP; Hambleton et al., 1991), expected a posteriori estimation (EAP; Bock & Mislevy, 1982),
and EAP summed scoring (Thissen et al., 1995; Thissen & Orlando, 2001).
}
\details{
For MAP scoring method, only the normal prior distribution is available for the population distribution.

When there are missing data in the response data set, the missing value must be specified in \code{missing}. The missing data are taken into account
when either of MLE, MAP, and EAP is used. However, there must be no missing data in the response data set when "EAP.SUM" is used.
One of possible ways to use "EAP.SUM" method when missing values exist is to remove rows with any missing values.

To speed up the ability estimation for MLE, MAP, and EAP methods, this function applies a parallel process using multiple logical CPU cores.
You can set the number of logical CPU cores by specifying a positive integer value in the argument \code{ncore}. Default value is 1.

Note that the standard errors of ability estimates are computed using observed information functions.
}
\examples{
## the use of a "-prm.txt" file obtained from a flexMIRT
flex_prm <- system.file("extdata", "flexmirt_sample-prm.txt", package = "irtplay")

# read item parameters and transform them to item meta data
x <- bring.flexmirt(file=flex_prm, "par")$Group1$full_df

# generate examinees abilities
set.seed(12)
theta <- rnorm(10)

# simulate the item response data
data <- simdat(x, theta, D=1)

\donttest{
# estimate the abilities using MLE
est_score(x, data, D=1, method="MLE", range=c(-4, 4), se=TRUE, ncore=2)

# estimate the abilities using MAP
est_score(x, data, D=1, method="MAP", norm.prior=c(0, 1), nquad=30, se=TRUE, ncore=2)

# estimate the abilities using EAP summed scoring
est_score(x, data, D=1, method="EAP.SUM", norm.prior=c(0, 1), nquad=30)
}

}
\references{
Bock, R. D., & Mislevy, R. J. (1982). Adaptive EAP estimation of ability in a microcomputer environment. \emph{Psychometrika, 35}, 179-198.

Hambleton, R. K., Swaminathan, H., & Rogers, H. J. (1991).\emph{Fundamentals of item response theory}. Newbury Park, CA: Sage.

Thissen, D. & Orlando, M. (2001). Item response theory for items scored in two categories. In D. Thissen & H. Wainer (Eds.),
\emph{Test scoring} (pp.73-140). Mahwah, NJ: Lawrence Erlbaum.

Thissen, D., Pommerich, M., Billeaud, K., & Williams, V. S. (1995). Item Response Theory
for Scores on Tests Including Polytomous Items with Ordered Responses. \emph{Applied Psychological
Measurement, 19}(1), 39-49.
}
\seealso{
\code{\link{irtfit}}, \code{\link{test.info}}, \code{\link{simdat}}, \code{\link{shape_df}}, \code{\link{gen.weight}}
}
\author{
Hwanggyu Lim \email{hglim83@gmail.com}
}
