% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdif2.R
\name{rdif2}
\alias{rdif2}
\title{IRT residual-based DIF (RDIF) detection framework2}
\usage{
rdif2(
  x,
  data,
  score = NULL,
  se = NULL,
  reg.cr = FALSE,
  group,
  focal.name,
  D = 1,
  alpha = 0.05,
  missing = NA,
  purify = FALSE,
  purify.by = c("rdif_rs", "rdif_r", "rdif_s"),
  max.iter = 10,
  min.resp = NULL,
  method = "MLE",
  range = c(-4, 4),
  norm.prior = c(0, 1),
  nquad = 41,
  weights = NULL,
  ncore = 1,
  verbose = TRUE
)
}
\arguments{
\item{x}{A data frame containing the item metadata (e.g., item parameters, number of categories, models ...), an object of class \code{\link{est_item}}
obtained from the function \code{\link{est_item}}, or an object of class \code{\link{est_irt}} obtained from the function \code{\link{est_irt}}.
The data frame of item metadata can be easily obtained using the function \code{\link{shape_df}}. See \code{\link{est_irt}}, \code{\link{irtfit}}, 
\code{\link{test.info}} or \code{\link{simdat}} for more details about the item metadata.}

\item{data}{A matrix containing examinees' response data for the items in the argument \code{x}. A row and column indicate
the examinees and items, respectively.}

\item{score}{A vector of examinees' ability estimates. If the abilities are not provided, \code{\link{rdif}} function estimates the abilities before 
computing RDIF statistics. See \code{\link{est_score}} for more details about scoring methods. Default is NULL.}

\item{group}{A numeric or character vector indicating group membership of examinees. The length of vector should the same with the number of rows 
in the response data matrix.}

\item{focal.name}{A single numeric or character indicating the level of group which corresponds to the focal group. 
For example, if \code{group = c(0, 1, 0, 1, 1)} and '1' indicates the focal group, then \code{focal.name = 1}.}

\item{D}{A scaling factor in IRT models to make the logistic function as close as possible to the normal ogive function (if set to 1.7).
Default is 1.}

\item{alpha}{A numeric value to specify significance \eqn{\alpha}-level of the hypothesis test using the RDIF fit statistics.
Default is .05.}

\item{missing}{A value indicating missing values in the response data set. Default is NA.}

\item{purify}{A logical value indicating whether a purification process will be implemented or not. Default is FALSE.}

\item{purify.by}{A character string specifying a RDIF statistic with which the purification is implemented. Available statistics 
are "rdif_rs" for \eqn{RDIF_{RS}}, "rdif_r" for \eqn{RDIF_{R}}, and "rdif_s" for \eqn{RDIF_{S}}.}

\item{max.iter}{An integer value specifying a maximum number of iterations for the purification process. Default is 10.}

\item{min.resp}{An integer value?? Default is NULL.}

\item{method}{A character string indicating a scoring method. Available methods are "MLE" for the maximum likelihood estimation, 
"MAP" for the maximum a posteriori estimation, and "EAP" for the expected a posteriori estimation. Default method is "MLE".}

\item{range}{A numeric vector of two components to restrict the range of ability scale for the MLE. Default is c(-4, 4).}

\item{norm.prior}{A numeric vector of two components specifying a mean and standard deviation of the normal prior distribution.
These two parameters are used to obtain the gaussian quadrature points and the corresponding weights from the normal distribution. Default is
c(0,1). Ignored if \code{method} is "MLE".}

\item{nquad}{An integer value specifying the number of gaussian quadrature points from the normal prior distribution. Default is 41.
Ignored if \code{method} is "MLE" or "MAP".}

\item{weights}{A two-column matrix or data frame containing the quadrature points (in the first column) and the corresponding weights
(in the second column) of the latent variable prior distribution. The weights and quadrature points can be easily obtained
using the function \code{\link{gen.weight}}. If NULL and \code{method} is "EAP", default values are used (see the arguments
of \code{norm.prior} and \code{nquad}). Ignored if \code{method} is "MLE" or "MAP".}

\item{ncore}{The number of logical CPU cores to use. Default is 1. See \code{\link{est_score}} for details.}

\item{verbose}{A logical value. If TRUE, the progress messages of purification procedure are suppressed. Default is TRUE.}
}
\description{
This function computes three RDIF statistics (Lim, Choe, Han, Lee, & Hong, 2021), 
which are \eqn{RDIF_{R}}, \eqn{RDIF_{S}}, and \eqn{RDIF_{RS}}, for each item. \eqn{RDIF_{R}} primarily 
captures the typical contrast in raw residual pattern between two groups caused by uniform DIF whereas 
\eqn{RDIF_{S}} primarily captures the typical contrast in squared residual pattern between two groups caused
by nonuniform DIF. \eqn{RDIF_{RS}} can reasonably capture both types of DIF.
}
