% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_flexmirt.R
\name{write.flexmirt}
\alias{write.flexmirt}
\title{Write a "-prm.txt" file for flexMIRT}
\usage{
write.flexmirt(x, file = NULL, norm.pop = c(0, 1), rePrm = TRUE)
}
\arguments{
\item{x}{A data.frame containing the item meta data (e.g., item parameters, number of categories, models ...).
See \code{\link{irtfit}}, \code{\link{test.info}}, or \code{\link{simdat}} for more details about the item meta data.
This data.frame can be easily obtained using the function \code{\link{shape_df}}.}

\item{file}{The destination file name.}

\item{norm.pop}{A numeric vector of two components specifying a mean and standard deviation of the normal
population distribution. Default is c(0,1).}

\item{rePrm}{A logical value indicating whether the item parameters in the item meta data
are the reparameterized item parameters. If TRUE, the item intercepts and logits of item guessing parameters
should be included in the item meta data. If FALSE, the item difficulty and item guessing parameters
should be included in the item meta data.}
}
\value{
A "-prm.txt" file.
}
\description{
This function writes an output file of "-prm.txt" for flexMIRT. The current version of this function can be used
only for the unidimensional IRT models.
}
\examples{
## use the simulated CAT data
# extract the item meta data
x <- simCAT_MX$item.prm

# set a name of "-prm.txt" file
temp_prm <- file.path(tempdir(), "temp-prm.txt")

# write out the "-prm.txt" file
write.flexmirt(x, file=temp_prm, norm.pop=c(0, 1), rePrm=FALSE)

}
\references{
Cai, L. (2015). flexMIRT 3.0 Flexible multilevel multidimensional item analysis and test scoring [Computer software].
Chapel Hill, NC: Vector Psychometric Group.
}
\author{
Hwanggyu Lim \email{hglim83@gmail.com}
}
